Manually made changes including the fix from #4 with my preferred coding style
One of my main complaints with the pull request was the splitting up of the LINQ statement into multiple statements. I definitely agree that it was too messy and nearly impossible to read or debug, so I kept the spirit of the pull request by moving the logic into separate methods... but still kept everything within a single statement.